README.md
The analysis focused on pitchers and past CY Young award winners. A logistic regression model was fitted, and linear discriminant analysis was applied using the LDA() function from the MASS library. It was found that longevity and consistency in a pitcher increased the probability of winning the CY Young award.